bind: considerably simplify binding creation
