+++ pkg/apis/kubebind/v1alpha1/helpers/crd.go
@@ -1,0 +1,42 @@
+/*
+Copyright 2022 The Kube Bind Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package helpers
+
+import (
+	"strings"
+
+	"k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/types"
+
+	"github.com/kube-bind/kube-bind/pkg/apis/kubebind/v1alpha1"
+)
+
+func IsOwnedByBinding(name string, uid types.UID, refs []v1.OwnerReference) bool {
+	for _, ref := range refs {
+		parts := strings.SplitN(ref.APIVersion, "/", 2)
+		if parts[0] != v1alpha1.SchemeGroupVersion.Group || ref.Kind != "APIServiceBinding" {
+			continue
+		}
+		if ref.Name != name {
+			continue
+		}
+		if uid == "" || ref.UID == uid {
+			return true
+		}
+	}
+	return false
+}
